# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Regulatory Tracking Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /regulations:
    get:
      operationId: listRegulations
      tags: [regulations]
      summary: List all regulations
      parameters:
        - name: jurisdiction
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of regulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Regulation"
    post:
      operationId: createRegulation
      tags: [regulations]
      summary: Create a new regulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegulationInput"
      responses:
        "201":
          description: Created regulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
  /regulations/{id}:
    get:
      operationId: getRegulation
      tags: [regulations]
      summary: Get a regulation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Regulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRegulation
      tags: [regulations]
      summary: Update a regulation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegulationInput"
      responses:
        "200":
          description: Updated regulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRegulation
      tags: [regulations]
      summary: Delete a regulation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Regulation:
      type: object
      properties:
        id:
          type: integer
        lawName:
          type: string
        jurisdiction:
          type: string
          enum: [federal, state, international]
        status:
          type: string
          enum: [proposed, enacted, effective]
        summary:
          type: string
        relevantPolicies:
          type: array
          items:
            type: string
        nextAction:
          type: string
        effectiveDate:
          type: string
          nullable: true
        deadlineDate:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - lawName
        - jurisdiction
        - status
        - summary
        - relevantPolicies
        - nextAction
        - createdAt
        - updatedAt
    CreateRegulationInput:
      type: object
      properties:
        lawName:
          type: string
        jurisdiction:
          type: string
          enum: [federal, state, international]
        status:
          type: string
          enum: [proposed, enacted, effective]
        summary:
          type: string
        relevantPolicies:
          type: array
          items:
            type: string
        nextAction:
          type: string
        effectiveDate:
          type: string
          nullable: true
        deadlineDate:
          type: string
          nullable: true
      required:
        - lawName
        - jurisdiction
        - status
        - summary
        - relevantPolicies
        - nextAction
...