# lib/api-zod/src/generated/types/createRegulationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRegulationInputJurisdiction } from "./createRegulationInputJurisdiction";
import type { CreateRegulationInputStatus } from "./createRegulationInputStatus";

export interface CreateRegulationInput {
  lawName: string;
  jurisdiction: CreateRegulationInputJurisdiction;
  status: CreateRegulationInputStatus;
  summary: string;
  relevantPolicies: string[];
  nextAction: string;
  effectiveDate?: string | null;
  deadlineDate?: string | null;
}