# lib/api-zod/src/generated/types/regulation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Regulatory Tracking Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { RegulationJurisdiction } from "./regulationJurisdiction";
import type { RegulationStatus } from "./regulationStatus";

export interface Regulation {
  id: number;
  lawName: string;
  jurisdiction: RegulationJurisdiction;
  status: RegulationStatus;
  summary: string;
  relevantPolicies: string[];
  nextAction: string;
  effectiveDate?: string | null;
  deadlineDate?: string | null;
  createdAt: string;
  updatedAt: string;
}